+++ components/http/okHttp/src/test/java/com/microsoft/kiota/http/middleware/ParametersNameDecodingHandlerTest.java
@@ -1,9 +1,9 @@
-package com.microsoft.kiota.http;
+package com.microsoft.kiota.http.middleware;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
 
-import com.microsoft.kiota.http.middleware.ParametersNameDecodingHandler;
+import com.microsoft.kiota.http.KiotaClientFactory;
 import com.microsoft.kiota.http.middleware.options.ParametersNameDecodingOption;
 
 import okhttp3.Interceptor;