+++ components/http/okHttp/src/test/java/com/microsoft/kiota/http/middleware/UserAgentHandlerTest.java
@@ -1,11 +1,10 @@
-package com.microsoft.kiota.http;
+package com.microsoft.kiota.http.middleware;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
 import static org.junit.jupiter.api.Assertions.assertNull;
 import static org.mockito.Mockito.*;
 
-import com.microsoft.kiota.http.middleware.UserAgentHandler;
 import com.microsoft.kiota.http.middleware.options.UserAgentHandlerOption;
 
 import okhttp3.Headers;
@@ -60,8 +59,8 @@ class UserAgentHandlerTest {
         final UserAgentHandler handler = new UserAgentHandler();
         final Request request = new Request.Builder().url("http://localhost").build();
         when(mockChain.request()).thenReturn(request);
+        handler.intercept(mockChain);
         Response response = handler.intercept(mockChain);
-        response = handler.intercept(mockChain);
         final Request result = response.request();
         assertNotNull(response);
         assertNotNull(result);